+++ .rubocop.yml
@@ -39,7 +39,10 @@ Style/Tab:
   Exclude:
     - 'spec/indent_spec.rb'
 
-Style/TrailingCommaInLiteral:
+Style/TrailingCommaInArrayLiteral:
+  Enabled: false
+
+Style/TrailingCommaInHashLiteral:
   Enabled: false
 
 Style/TrailingCommaInArguments:
