+++ .rubocop_todo.yml
@@ -1,28 +1,261 @@
-Metrics/AbcSize:
-  Enabled: false
+# This configuration was generated by
+# `rubocop --auto-gen-config`
+# on 2019-03-09 17:41:12 +0300 using RuboCop version 0.63.0.
+# The point is for the user to remove these configuration records
+# one by one as the offenses are removed from the code base.
+# Note that changes in the inspected code, or installation of new
+# versions of RuboCop, may require this file to be generated again.
 
-Metrics/ClassLength:
-  Enabled: false
-
-Metrics/CyclomaticComplexity:
-  Enabled: false
-
-Metrics/LineLength:
-  Enabled: false
-
-Metrics/MethodLength:
-  Enabled: false
-
-Metrics/ModuleLength:
-  Enabled: false
-
-Metrics/PerceivedComplexity:
-  Enabled: false
-
-Style/Documentation:
-  Enabled: false
-
-Style/PredicateName:
+# Offense count: 1
+# Configuration parameters: Include.
+# Include: **/*.gemspec
+Gemspec/RequiredRubyVersion:
   Exclude:
-    - 'lib/hamdown_parser/line_parser.rb'
-    - 'lib/hamdown_parser/ruby_multiline.rb'
+    - 'haml_parser.gemspec'
+
+# Offense count: 5
+# Cop supports --auto-correct.
+Layout/EmptyLineAfterGuardClause:
+  Exclude:
+    - 'lib/hamdown_core/element_parser.rb'
+    - 'lib/hamdown_core/indent_tracker.rb'
+
+# Offense count: 31
+# Cop supports --auto-correct.
+Layout/EmptyLineAfterMagicComment:
+  Enabled: false
+
+# Offense count: 38
+# Cop supports --auto-correct.
+# Configuration parameters: EnforcedStyle.
+# SupportedStyles: auto_detection, squiggly, active_support, powerpack, unindent
+Layout/IndentHeredoc:
+  Exclude:
+    - 'spec/attribute_spec.rb'
+    - 'spec/comment_spec.rb'
+    - 'spec/doctype_spec.rb'
+    - 'spec/element_spec.rb'
+    - 'spec/filter_spec.rb'
+    - 'spec/haml_comment_spec.rb'
+    - 'spec/html_style_attribute_spec.rb'
+    - 'spec/indent_spec.rb'
+    - 'spec/multiline_spec.rb'
+    - 'spec/plain_spec.rb'
+    - 'spec/script_spec.rb'
+    - 'spec/silent_script_spec.rb'
+    - 'spec/whitespace_spec.rb'
+
+# Offense count: 1
+# Cop supports --auto-correct.
+Layout/SpaceAfterComma:
+  Exclude:
+    - 'lib/hamdown_compiler.rb'
+
+# Offense count: 1
+# Cop supports --auto-correct.
+# Configuration parameters: EnforcedStyle.
+# SupportedStyles: final_newline, final_blank_line
+Layout/TrailingBlankLines:
+  Exclude:
+    - 'lib/hamdown_compiler.rb'
+
+# Offense count: 11
+Lint/InterpolationCheck:
+  Exclude:
+    - 'spec/attribute_spec.rb'
+    - 'spec/element_spec.rb'
+    - 'spec/html_style_attribute_spec.rb'
+    - 'spec/script_spec.rb'
+
+# Offense count: 1
+# Cop supports --auto-correct.
+Lint/ScriptPermission:
+  Exclude:
+    - 'benchmark/parse.rb'
+
+# Offense count: 2
+Lint/ShadowingOuterLocalVariable:
+  Exclude:
+    - 'lib/hamdown_compiler.rb'
+
+# Offense count: 1
+Lint/UselessAssignment:
+  Exclude:
+    - 'lib/hamdown_compiler.rb'
+
+# Offense count: 13
+Metrics/AbcSize:
+  Max: 80
+
+# Offense count: 20
+# Configuration parameters: CountComments, ExcludedMethods.
+# ExcludedMethods: refine
+Metrics/BlockLength:
+  Max: 171
+
+# Offense count: 2
+# Configuration parameters: CountComments.
+Metrics/ClassLength:
+  Max: 247
+
+# Offense count: 7
+Metrics/CyclomaticComplexity:
+  Max: 17
+
+# Offense count: 21
+# Configuration parameters: CountComments, ExcludedMethods.
+Metrics/MethodLength:
+  Max: 67
+
+# Offense count: 1
+# Configuration parameters: CountComments.
+Metrics/ModuleLength:
+  Max: 129
+
+# Offense count: 3
+Metrics/PerceivedComplexity:
+  Max: 13
+
+# Offense count: 3
+# Configuration parameters: NamePrefix, NamePrefixBlacklist, NameWhitelist, MethodDefinitionMacros.
+# NamePrefix: is_, has_, have_
+# NamePrefixBlacklist: is_, has_, have_
+# NameWhitelist: is_a?
+# MethodDefinitionMacros: define_method, define_singleton_method
+Naming/PredicateName:
+  Exclude:
+    - 'spec/**/*'
+    - 'lib/hamdown_core/line_parser.rb'
+    - 'lib/hamdown_core/ruby_multiline.rb'
+
+# Offense count: 1
+# Configuration parameters: MinNameLength, AllowNamesEndingInNumbers, AllowedNames, ForbiddenNames.
+# AllowedNames: io, id, to, by, on, in, at, ip, db
+Naming/UncommunicativeMethodParamName:
+  Exclude:
+    - 'lib/hamdown_core/parser.rb'
+
+# Offense count: 3
+# Cop supports --auto-correct.
+Performance/StringReplacement:
+  Exclude:
+    - 'lib/hamdown_compiler.rb'
+
+# Offense count: 3
+# Cop supports --auto-correct.
+# Configuration parameters: EnforcedStyle, SingleLineConditionsOnly, IncludeTernaryExpressions.
+# SupportedStyles: assign_to_condition, assign_inside_condition
+Style/ConditionalAssignment:
+  Exclude:
+    - 'lib/hamdown_compiler.rb'
+    - 'lib/hamdown_core/parser.rb'
+
+# Offense count: 17
+Style/Documentation:
+  Exclude:
+    - 'spec/**/*'
+    - 'test/**/*'
+    - 'lib/hamdown_compiler.rb'
+    - 'lib/hamdown_core.rb'
+    - 'lib/hamdown_core/ast.rb'
+    - 'lib/hamdown_core/cli.rb'
+    - 'lib/hamdown_core/element_parser.rb'
+    - 'lib/hamdown_core/error.rb'
+    - 'lib/hamdown_core/filter_parser.rb'
+    - 'lib/hamdown_core/indent_tracker.rb'
+    - 'lib/hamdown_core/line_parser.rb'
+    - 'lib/hamdown_core/parser.rb'
+    - 'lib/hamdown_core/ruby_multiline.rb'
+    - 'lib/hamdown_core/script_parser.rb'
+    - 'lib/hamdown_core/utils.rb'
+    - 'lib/hamdown_transformer.rb'
+
+# Offense count: 1
+# Cop supports --auto-correct.
+Style/Encoding:
+  Exclude:
+    - 'haml_parser.gemspec'
+
+# Offense count: 1
+# Cop supports --auto-correct.
+Style/ExpandPathArguments:
+  Exclude:
+    - 'haml_parser.gemspec'
+
+# Offense count: 6
+# Cop supports --auto-correct.
+Style/MutableConstant:
+  Exclude:
+    - 'lib/hamdown_core/element_parser.rb'
+    - 'lib/hamdown_core/line_parser.rb'
+    - 'lib/hamdown_core/parser.rb'
+
+# Offense count: 3
+# Cop supports --auto-correct.
+# Configuration parameters: EnforcedStyle.
+# SupportedStyles: both, prefix, postfix
+Style/NegatedIf:
+  Exclude:
+    - 'lib/hamdown_compiler.rb'
+
+# Offense count: 11
+# Cop supports --auto-correct.
+# Configuration parameters: AutoCorrect, EnforcedStyle, IgnoredMethods.
+# SupportedStyles: predicate, comparison
+Style/NumericPredicate:
+  Exclude:
+    - 'spec/**/*'
+    - 'lib/hamdown_compiler.rb'
+    - 'lib/hamdown_core/element_parser.rb'
+    - 'lib/hamdown_core/parser.rb'
+    - 'lib/hamdown_core/utils.rb'
+    - 'lib/hamdown_transformer.rb'
+
+# Offense count: 2
+# Cop supports --auto-correct.
+# Configuration parameters: ConvertCodeThatCanStartToReturnNil, Whitelist.
+# Whitelist: present?, blank?, presence, try, try!
+Style/SafeNavigation:
+  Exclude:
+    - 'lib/hamdown_core/ast.rb'
+    - 'lib/hamdown_core/script_parser.rb'
+
+# Offense count: 2
+# Cop supports --auto-correct.
+# Configuration parameters: EnforcedStyle, ConsistentQuotesInMultiline.
+# SupportedStyles: single_quotes, double_quotes
+Style/StringLiterals:
+  Exclude:
+    - 'lib/hamdown_compiler.rb'
+
+# Offense count: 1
+# Cop supports --auto-correct.
+# Configuration parameters: MinSize.
+# SupportedStyles: percent, brackets
+Style/SymbolArray:
+  EnforcedStyle: brackets
+
+# Offense count: 1
+# Cop supports --auto-correct.
+Style/UnneededInterpolation:
+  Exclude:
+    - 'lib/hamdown_compiler.rb'
+
+# Offense count: 1
+# Cop supports --auto-correct.
+Style/WhileUntilModifier:
+  Exclude:
+    - 'lib/hamdown_core/parser.rb'
+
+# Offense count: 6
+# Cop supports --auto-correct.
+Style/ZeroLengthPredicate:
+  Exclude:
+    - 'lib/hamdown_compiler.rb'
+    - 'lib/hamdown_transformer.rb'
+
+# Offense count: 31
+# Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
+# URISchemes: http, https
+Metrics/LineLength:
+  Max: 168
