+++ Gemfile
@@ -1,8 +1,5 @@
 # frozen_string_literal: true
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in hamdown_parser.gemspec
+# Specify your gem's dependencies in hamdown_core.gemspec
 gemspec
-
-gem 'solargraph'
-gem 'rubocop'
