+++ bin/hamdown_transformer
@@ -1,0 +1,12 @@
+#!/usr/bin/env ruby
+
+# frozen_string_literal: true
+
+require_relative '../lib/hamdown_core/parser'
+require_relative '../lib/hamdown_core/transformer'
+require 'pp'
+
+parser = HamdownCore::Parser.new(filename: 'input.haml')
+ast = parser.call(File.read('input.haml'))
+transformed_ast = HamdownCore::Transformer.call(ast)
+pp transformed_ast
