+++ exe/hamdown_core
@@ -1,0 +1,7 @@
+#!/usr/bin/env ruby
+
+# frozen_string_literal: true
+
+require './lib/hamdown_core'
+
+HamdownCore::Cli.call(ARGV)
