--- exe/hamdown_parser
@@ -1,11 +1,0 @@
-#!/usr/bin/env ruby
-
-# require 'hamdown_parser/cli'
-# HamdownParser::CLI.start(ARGV)
-
-require_relative '../lib/hamdown_parser/parser'
-require 'pp'
-
-parser = HamdownParser::Parser.new(filename: 'input.haml')
-ast = parser.call(File.read('input.haml'))
-pp ast
