--- exe/hamdown_transformer
@@ -1,10 +1,0 @@
-#!/usr/bin/env ruby
-
-require_relative '../lib/hamdown_parser/parser'
-require_relative '../lib/hamdown_transformer'
-require 'pp'
-
-parser = HamdownParser::Parser.new(filename: 'input.haml')
-ast = parser.call(File.read('input.haml'))
-transformed_ast = HamdownTransformer.call(ast)
-pp transformed_ast
