+++ lib/hamdown_core.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+require_relative 'hamdown_core/version'
+require_relative 'hamdown_core/cli'
+require_relative 'hamdown_core/engine'
+
+module HamdownCore
+  # Your code goes here...
+end
