+++ lib/hamdown_core/utils.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
-module HamdownParser
+module HamdownCore
   module Utils
     module_function
