+++ lib/hamdown_core/version.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+module HamdownCore
+  VERSION = '0.5.0'
+end
