--- lib/hamdown_parser.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-require_relative 'hamdown_parser/version'
-
-module HamdownParser
-  # Your code goes here...
-end
