--- lib/hamdown_parser/version.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-module HamdownParser
-  VERSION = '0.4.1'
-end
