--- lib/hamdown_transformer.rb
@@ -1,50 +1,0 @@
-# frozen_string_literal: true
-
-module HamdownTransformer
-  class << self
-    def call(ast)
-      transform(ast)
-    end
-
-    private
-
-    def transform(root_node)
-      collected_nodes = []
-      new_root_node = root_node.clone
-      new_root_node.children = []
-      root_node.children.each do |node|
-        if node.respond_to?(:markdownable?) && node.markdownable?
-          collected_nodes << node
-          next
-        end
-
-        if collected_nodes.size > 0
-          filter = create_filter(collected_nodes)
-          collected_nodes = []
-          new_root_node << filter
-        end
-
-        if node.respond_to?(:children) && node.children.size > 0
-          node = transform(node)
-        end
-
-        new_root_node << node
-      end
-      if collected_nodes.size > 0
-        filter = create_filter(collected_nodes)
-        collected_nodes = []
-        new_root_node << filter
-      end
-      new_root_node
-    end
-
-    def create_filter(nodes)
-      filter = HamdownParser::Ast::Filter.new
-      filter.name = 'markdown'
-      nodes.each do |collected_node|
-        filter.texts << collected_node.text
-      end
-      filter
-    end
-  end
-end
